pyproject.toml
[tool.poetry]
name = "psp"
version = "0.1.0"
description = "Photovoltaic per site modeling"
authors = ["Simon Lemieux <1105380+simlmx@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
scikit-learn = "^1.1.3"
pvlib = "^0.9.3"
numpy = "^1.23.5"
torch = "^1.13.1"
torchdata = "^0.5.0"
xarray = "^2022.12.0"
click = "^8.1.3"
google-cloud-storage = "^2.7.0"
zarr = "^2.13.3"
fsspec = "^2022.11.0"
gcsfs = "^2022.11.0"
dask = "^2022.12.1"
h5netcdf = "^1.1.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
altair = "^4.2.0"
plotly = "^5.11.0"
astral = "^3.2"
pytest = "^7.2.0"
flake8 = "^6.0.0"
isort = "^5.10.1"
jupyter-contrib-nbextensions = "^0.7.0"
black = {extras = ["jupyter"], version = "^22.10.0"}
scipy = "^1.9.3"
tqdm = "^4.64.1"
pandas-stubs = "^1.5.2.221213"
types-tqdm = "^4.64.7.9"
mypy = "^0.991"
pyarrow = "^10.0.1"
types-pyyaml = "^6.0.12.2"
pylint = "^2.15.9"
pydocstyle = "^6.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.mypy]
check_untyped_defs = true
# Too many packages have missing stubs.
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = 'astral.*'
ignore_errors = true

[tool.pydocstyle]
# In particular ignore all the "missing docstring" errors.
ignore = 'D10,D203,D205,D212,D213,D400,D413,D415,D417'
